Refactor product endpoints into dedicated API router
